Fix broken parallelism in newly supported jest-circus
