+++ detox/runners/jest/specReporter.js
@@ -1,9 +1,9 @@
 const argparse = require('../../src/utils/argparse');
+const runnerInfo = require('./runnerInfo');
 
 if (argparse.getArgValue('reportSpecs') === 'true') {
-  const runnerInfo = require('./runnerInfo');
   const Reporter = runnerInfo.isJestCircus ? require('./SpecReporterCircus') : require('./SpecReporterJasmine');
   module.exports = new Reporter();
 } else {
-  module.exports = {};
+  module.exports = runnerInfo.isJestCircus ? require('./CircusTestEventListenerBase').stubEventsListener : {};
 }
