+++ examples/demo-react-native-jest/e2e/app-goodbye.test.js
@@ -1,0 +1,10 @@
+describe('Example (goodbye)', () => {
+  beforeEach(async () => {
+    await device.reloadReactNative();
+  });
+
+  it('should show goodbye screen after tap', async () => {
+    await element(by.id('goodbye_button')).tap();
+    await expect(element(by.text('Goodbye, World!!!'))).toBeVisible();
+  });
+});
