+++ examples/demo-react-native-jest/e2e/app-hello.test.js
@@ -1,17 +1,17 @@
-describe('Example', () => {
+describe('Example (hello)', () => {
   beforeEach(async () => {
     await device.reloadReactNative();
   });
-  
+
   it('should have welcome screen', async () => {
     await expect(element(by.id('welcome'))).toBeVisible();
   });
-  
+
   it('should show hello screen after tap', async () => {
     await element(by.id('hello_button')).tap();
     await expect(element(by.text('Hello!!!'))).toBeVisible();
   });
-  
+
   it('should show world screen after tap', async () => {
     await element(by.id('world_button')).tap();
     await expect(element(by.text('World!!!'))).toBeVisible();
