+++ examples/demo-react-native-jest/package.json
@@ -4,9 +4,13 @@
   "private": true,
   "scripts": {
     "test:ios-release": "detox test --configuration ios.sim.release -l verbose",
+    "test:ios-release-ci": "detox test --configuration ios.sim.release -l verbose --workers 2",
     "test:jest-circus:ios-release": "detox test --configuration ios.sim.release -l verbose -o e2e/config-circus.json",
+    "test:jest-circus:ios-release-ci": "detox test --configuration ios.sim.release -l verbose -o e2e/config-circus.json --workers 2",
     "test:android-release": "detox test --configuration android.emu.release -l verbose",
-    "test:jest-circus:android-release": "detox test --configuration android.emu.release -l verbose -o e2e/config-circus.json"
+    "test:android-release-ci": "detox test --configuration android.emu.release -l verbose --workers 2",
+    "test:jest-circus:android-release": "detox test --configuration android.emu.release -l verbose -o e2e/config-circus.json",
+    "test:jest-circus:android-release-ci": "detox test --configuration android.emu.release -l verbose -o e2e/config-circus.json --workers 2"
   },
   "devDependencies": {
     "detox": "^14.3.0",
