+++ scripts/demo-projects.android.sh
@@ -16,6 +16,6 @@ run_f "npm run test:android-explicit-require -- --headless"
 popd
 
 pushd examples/demo-react-native-jest
-run_f "npm run test:android-release -- --headless"
-run_f "npm run test:jest-circus:android-release -- --headless"
+run_f "npm run test:android-release-ci -- --headless"
+run_f "npm run test:jest-circus:android-release-ci -- --headless"
 popd
